+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Language property to template locals.
+- Ability to omit template file extension.
 
 ## [0.3.0] - 2019-06-21
 ### Added